+++ api-tests/relationships/crud/one-to-one.test.js
@@ -110,6 +110,27 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
           });
         }
 
+        describe('Count', () => {
+          test(
+            'Count',
+            runner(setupKeystone, async ({ keystone }) => {
+              await createInitialData(keystone);
+              const { data, errors } = await graphqlRequest({
+                keystone,
+                query: `
+                {
+                  _allCompaniesMeta { count }
+                  _allLocationsMeta { count }
+                }
+            `,
+              });
+              expect(errors).toBe(undefined);
+              expect(data._allCompaniesMeta.count).toEqual(3);
+              expect(data._allLocationsMeta.count).toEqual(3);
+            })
+          );
+        });
+
         describe('Create', () => {
           test(
             'With connect',
@@ -168,7 +189,7 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
             })
           );
 
-          test.failing(
+          test(
             'With nested connect',
             runner(setupKeystone, async ({ keystone }) => {
               const { companies } = await createInitialData(keystone);
@@ -215,7 +236,7 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
             })
           );
 
-          test.failing(
+          test(
             'With nested create',
             runner(setupKeystone, async ({ keystone }) => {
               const locationName = sampleOne(alphanumGenerator);